refactor: tag 관련 api 함수 호출 경로 수정
