+++ frontend/techpick/src/apis/tag/createTag.ts
@@ -1,10 +1,14 @@
-import { apiClient } from '@/apis';
+import { apiClient } from '../apiClient';
+import { API_URLS } from '../apiConstants';
 import { CreateTagRequestType, CreateTagResponseType } from '@/types';
 
 export const createTag = async (createTag: CreateTagRequestType) => {
-  const response = await apiClient.post<CreateTagResponseType>('tag', {
-    body: JSON.stringify(createTag),
-  });
+  const response = await apiClient.post<CreateTagResponseType>(
+    API_URLS.CREATE_TAGS,
+    {
+      body: JSON.stringify(createTag),
+    }
+  );
   const data = await response.json();
 
   return data;
