+++ frontend/techpick/src/apis/tag/deleteTag.ts
@@ -1,4 +1,5 @@
-import { apiClient } from '@/apis';
+import { apiClient } from '../apiClient';
+import { API_URLS } from '../apiConstants';
 import { DeleteTagRequestType } from '@/types';
 
 export const deleteTag = async (tagId: DeleteTagRequestType['id']) => {
@@ -6,7 +7,7 @@ export const deleteTag = async (tagId: DeleteTagRequestType['id']) => {
     id: tagId,
   };
 
-  await apiClient.delete<never>(`tags`, {
+  await apiClient.delete<never>(API_URLS.DELETE_TAGS, {
     json: deleteTagRequest,
   });
 };
